* 7.0:
  always pass microseconds to usleep as integers
  Bump Symfony version to 7.0.2
  Update VERSION for 7.0.1
  Update CHANGELOG for 7.0.1
  Bump Symfony version to 6.4.2
  Update VERSION for 6.4.1
  Update CHANGELOG for 6.4.1
  Bump Symfony version to 6.3.11
  Update VERSION for 6.3.10
  Update CHANGELOG for 6.3.10
  Bump Symfony version to 5.4.34
  Update VERSION for 5.4.33
  Update CONTRIBUTORS for 5.4.33
  Update CHANGELOG for 5.4.33

+++ src/Symfony/Component/Mailer/Transport/AbstractTransport.php
@@ -129,7 +129,7 @@ abstract class AbstractTransport implements TransportInterface
         $sleep = (1 / $this->rate) - (microtime(true) - $this->lastSent);
         if (0 < $sleep) {
             $this->logger->debug(sprintf('Email transport "%s" sleeps for %.2f seconds', __CLASS__, $sleep));
-            usleep($sleep * 1000000);
+            usleep((int) ($sleep * 1000000));
         }
         $this->lastSent = microtime(true);
     }
